requirements.txt
SQLAlchemy==2.0.16
configparser==5.3.0
fastapi==0.97.0
psycopg2-binary==2.9.6
pydantic==1.10.9
pymysql==1.0.3
PyPDF2==3.0.1
uvicorn==0.22.0
requests~=2.31.0